change float to double

+++ circuitSimCuSolverDn.cu
@@ -26,8 +26,8 @@ int main(int argc, char *argv[]) {
 
     int augmented_matrix_dim=get_Matrix_Dim_from_nodes(elementList,elementListLength);
     printf("%d\n",augmented_matrix_dim);
-    std::vector<float> conductance(augmented_matrix_dim*augmented_matrix_dim);
-    std::vector<float> currents(augmented_matrix_dim);
+    std::vector<double> conductance(augmented_matrix_dim*augmented_matrix_dim);
+    std::vector<double> currents(augmented_matrix_dim);
     elementList_to_augmented_Matrix(elementList, elementListLength, conductance, currents, augmented_matrix_dim);
     for(int i=0;i<augmented_matrix_dim;i++)
     {
@@ -42,8 +42,8 @@ int main(int argc, char *argv[]) {
         printf("%f\n",currents[i]);
     }
     int matrix_dim=augmented_matrix_dim-1;
-    std::vector<float> conductance_definite((augmented_matrix_dim-1)*(augmented_matrix_dim-1));
-    std::vector<float> currents_definite(augmented_matrix_dim-1);
+    std::vector<double> conductance_definite((augmented_matrix_dim-1)*(augmented_matrix_dim-1));
+    std::vector<double> currents_definite(augmented_matrix_dim-1);
     augmented_Matrix_to_definite_matrix( elementListLength,  conductance,  currents,  conductance_definite, currents_definite,  augmented_matrix_dim);
     for(int i=0;i<matrix_dim;i++)
     {
@@ -53,13 +53,13 @@ int main(int argc, char *argv[]) {
         }
         printf("\n");
     }
-    std::vector<float> A=conductance_definite;
-    std::vector<float> B=currents_definite;
+    std::vector<double> A=conductance_definite;
+    std::vector<double> B=currents_definite;
 
     cusolverDnHandle_t cusolverH = NULL;
     cudaStream_t stream = NULL;
 
-    using data_type = float;
+    using data_type = double;
 
     const int64_t m = matrix_dim;
     const int64_t lda = m;
